Moved readOutputToken to StoneOutput token
